lib/fs/lockfile
A lockfile implementation that uses symbolic links.
The lock will point to /proc/<pid> and is therefore weak agains pid reuse.
This is only an issue if (a) a lock is not removed at exit and/or
(b) this module is bypassed, e.g. by calling rm -f <lock> directly.