lib/fs/disk: waitfor_disk(), get_disk(), do_fsck()
This module offers functions for processing disk device requests.
It's able to find a device by label, uuid or filesystem path and
to fsck it.
Additionally, it provides the ability to wait for a device to become
available.

Meant for usage in an initramfs init script,
but could have different use cases as well.