scripts/lib: add pack_system
A lib that allows to create image files (tarballs, squashfs files) of systems,
e.g. chroots.
See in-code documentation for usage, especially the file header comment.

This file is not part of module library (lib/) because its function names are
too generic (e.g. next(), which starts the declaration of a pack target).