add shlib build system
This commit adds a new build system, which will replace the current
one (make_scripts.sh, generate_script.sh,...) in future.

It mostly resides in lib/devel/shlib and provides functions for building
various types of scripts (shlib-linked, standalone, splitlib,
difference-standalone, simple copy, ...) as well as basic script verificiation
(syntax checking via $SHELL -n).

The frontend to these build modules is dobuild.sh (currently), which loads
and executes build instruction files (so-called recipes), which are simple
shell files. The result of running the dobuild frontend is a directory
containing scripts which can then be transferred to its destdir or packed
as tarball.

The new build system lacks documentation, which is TODO.
In the meantime, see files/recipe/ for examples.