lib/fs: add packlib/ modules (experimental)
This module is based on lib/fs/pack and scripts/lib/pack_system, and will
replace them in time.

fs/packlib offers functions for creating "pack targets" (which are functions
named pack_target_<name>()) and processing them.
These pack targets declare a directory that should be packed (relative to a
configurable root dir), a list of files/dirs that should be excluded and
a "pack type", which sets the backend used for packing the target. tar and
squashfs are supported for now.

In contrast to lib/fs/pack, packlib is multi-backend aware.
scripts/lib/pack_system tried to circumvent that by duplicating code etc.